[zutils] E: Move commands to subdirectory
Move standalone commands to subdirectory for the sake of sanity.

[zutils] B: copilot comments

Also update gitignores
