+++ examples/bin-hello/.nanvix/.gitignore
@@ -1,7 +1,9 @@
-venv/
-cache/
-sysroot/
-buildroot/
+__pycache__
+venv
+cache
+sysroot
+buildroot
+.yamllint.yml
+black.toml
 env.json
-nanvix.lock
-__pycache__/
+pyrightconfig.json
