+++ examples/lib-hello/.nanvix/.gitignore
@@ -1,8 +1,8 @@
-__pycache__/
-venv/
-cache/
-sysroot/
-buildroot/
+__pycache__
+venv
+cache
+sysroot
+buildroot
 .yamllint.yml
 black.toml
 env.json
