+++ pyproject.toml
@@ -29,7 +29,7 @@ dev = [
 
 [project.scripts]
 nanvix-zutil = "nanvix_zutil.__main__:main"
-nanvix-info = "nanvix_zutil.info:main"
+nanvix-info = "nanvix_zutil.commands.info:main"
 
 [project.urls]
 Homepage = "https://github.com/nanvix/zutils"
