+++ src/nanvix_zutil/commands/distclean.py
