+++ src/nanvix_zutil/commands/info.py
